[Test] Consolidate testHelpers
